% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.agroclimatology_list}
\alias{.agroclimatology_list}
\title{Agroclimatology List}
\usage{
.agroclimatology_list(file_list, isd_history, years)
}
\arguments{
\item{isd_history}{isd_history file from NCEI}

\item{years}{Years being requested}

\item{x}{A \code{data.table} of GSOD data from .download_data}
}
\value{
A list of GSOD stations suitable for agroclimatology work.
}
\description{
Agroclimatology List
}
\keyword{internal}
