% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.apply_process_csv}
\alias{.apply_process_csv}
\title{Process .gz files}
\usage{
.apply_process_csv(file_list, isd_history)
}
\arguments{
\item{file_list}{List of GSOD files}

\item{isd_history}{isd_history.csv file from NCEI provided by GSODR}
}
\value{
A \code{data.table} of GSOD weather data.
}
\description{
Internal function: Process .gz files
}
\keyword{internal}
