% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_url_exists}
\alias{.check_url_exists}
\title{Check That a URL Exists Before Downloading}
\usage{
.check_url_exists(x)
}
\arguments{
\item{x}{a URL for checking}
}
\value{
A numeric value representing the HTTP response.
}
\description{
Check That a URL Exists Before Downloading
}
\keyword{internal}
