% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.download_files}
\alias{.download_files}
\title{Download GSOD files from NCEI server}
\usage{
.download_files(station, years)
}
\arguments{
\item{station}{Station ID being requested. Optional}

\item{years}{Years being requested. Mandatory}
}
\value{
A list of data for processing before returning to user.
}
\description{
Internal function: Download GSOD files from NCEI server
}
