% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.subset_country_list}
\alias{.subset_country_list}
\title{Subset country list}
\usage{
.subset_country_list(country, file_list, isd_history, years)
}
\arguments{
\item{country}{Country of interest to subset on}

\item{isd_history}{isd_history.csv file from NCEI provided by GSODR}

\item{years}{Years being requested}

\item{GSOD_list}{List of GSOD files to be subset}
}
\value{
A list of stations in the requested country.
}
\description{
Subset country list
}
\keyword{internal}
\keyword{internal}
