% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.untar_files}
\alias{.untar_files}
\title{Untar GSOD Tar Archive Files}
\usage{
.untar_files(tar_files)
}
\arguments{
\item{tar_files}{a list of tar files located in in \code{tempdir()}}
}
\value{
Called for it's side-effects, untars the archive files in the
\code{tempdir()}.
}
\description{
Internal function: Untar GSOD Tar Archive Files
}
