% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_country}
\alias{.validate_country}
\title{Validate country requests}
\usage{
.validate_country(country, isd_history)
}
\arguments{
\item{country}{User requested country name}

\item{isd_history}{Data provided from NCEI on stations' locations and years}
}
\value{
A validated country name.
}
\description{
Internal function: Validate country requests
}
