% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_missing_days}
\alias{.validate_missing_days}
\title{Validate data for missing days}
\usage{
.validate_missing_days(max_missing, file_list)
}
\arguments{
\item{max_missing}{User entered maximum permissible missing days}

\item{GSOD_list}{A list of GSOD files that have been downloaded from NCEI}
}
\value{
A validated \code{list()} of GSOD files that meet requirements for
missing days.
}
\description{
Internal function: Validate data for missing days
}
