% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_station_data_years}
\alias{.validate_station_data_years}
\title{Validate Station Data for Years Available}
\usage{
.validate_station_data_years(station, isd_history, years)
}
\arguments{
\item{station}{User entered station ID}

\item{isd_history}{isd_history.csv from NCEI provided by GSODR}

\item{years}{User entered years for query}
}
\value{
\code{station_id} value, "station", \code{NA} if no match with available
data.
}
\description{
Internal function: Validate Station Data for Years Available
}
