% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_station_id}
\alias{.validate_station_id}
\title{Validate Station IDs}
\usage{
.validate_station_id(station, isd_history)
}
\arguments{
\item{station}{User entered station ID}

\item{isd_history}{isd_history.csv from NCEI provided by GSODR}
}
\value{
None unless an error with the years or invalid station ID.
}
\description{
Validate Station IDs
}
\keyword{internal}
