% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.validate_years}
\alias{.validate_years}
\title{Validate Years}
\usage{
.validate_years(years)
}
\arguments{
\item{years}{User entered years for request}
}
\value{
None unless error in years being requested by users.
}
\description{
Validate Years
}
\keyword{internal}
