% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_station_list.R
\name{fetch_station_list}
\alias{fetch_station_list}
\title{Download the Latest Station List From the NCDC Server}
\usage{
fetch_station_list()
}
\description{
This function downloads the latest station list from the NCDC FTP server.
This list includes metadata for all stations including unique identifiers,
country, state (if in US), latitude, longitude, elevation and when weather
observations begin and end.
}
\examples{
\dontrun{
GSOD_stations <- fetch_station_list()
}
}
\author{
Adam H Sparks \email{adamhsparks@gmail.com}
}

