% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_GSOD.R
\name{get_GSOD}
\alias{get_GSOD}
\title{Download, Clean and Generate New Variables From GSOD Weather Data}
\usage{
get_GSOD(start_year, end_year, max_missing)
}
\arguments{
\item{start_year}{The first year of the series of weather data to download}

\item{end_year}{The last year of the series of weather data to download}

\item{max_missing}{The maximum number of days allowed to be missing from a
station's data before it is excluded from .csv file output}
}
\description{
This function automates downloading and cleaning data from the
Global Summary of the Day (GSOD) data provided by the US National Climatic
Data Center (NCDC). Stations are individually checked for number of missing
days to assure data quality, stations with too many missing observations are
omitted. All units are converted to metric, e.g. feet to metres and
Fahrenheit to Celcius. Output is saved as a .csv file summarizing each year by
station, which includes vapor pressure and relative humidity variables
calculated from existing data in GSOD.

Be sure to have disk space free and allocate the proper time for this to run.
This is a time, processor and disk space intensive process.

For more information see the description of the data provided by NCDC,
\url{http://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt}
}
\details{
This function generates a GSOD_TPYYYY_XY.csv file in the respective
year directory containing the following data:
weather variables as columns
STNID - Station ID,
LAT - latitude,
LON - longitude,
ELEV.M - elevation in metres,
YEARMODA - Date in YYYY-MM-DD format,
YEAR - Year,
MONTH - Month,
DAY - Day,
TEMPC - Mean daily temperature in *C,
DEWPC-  Mean daily dewpoint in *C,
WDSPC - Mean daily wind speed value,
MAXC - Daily maximum temperature,
MINC - Daily minimum temperature,
ea - Mean daily actual vapor, pressure,
es - Mean daily saturation vapor pressure,
RH - Mean daily relative humidity
}
\examples{
# Download data for years 2009 and 2010 and generate yearly summary files,
# GSOD_TP2009_XY and GSOD_TP2010_XY files in folders 2009 and 2010 of your
# working directory with a maximum of five missing days per weather station
# allowed.

get_GSOD(start_year = 2009, end_year = 2010, max_missing = 5)
}

