% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory.R
\name{get_inventory}
\alias{get_inventory}
\title{Download and return a tidy data.frame of GSOD weather station data inventories}
\usage{
get_inventory()
}
\value{
\code{\link[base]{data.frame}} object of station inventories
}
\description{
The NCEI maintains a document,
<ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-inventory.txt>, which shows the
number of weather observations by station-year-month from the beginning of
the stations' records. This function retrieves that document, prints the
header to display the last update time and caclulates the percent
monthly coverage for each station reported.
}
\note{
The GSOD data, which are downloaded and manipulated by this R package,
stipulate that the following notice should be given.  \dQuote{The following
data and products may have conditions placed on their international
commercial use.  They can be used within the U.S. or for non-commercial
international activities without restriction.  The non-U.S. data cannot be
redistributed for commercial purposes.  Re-distribution of these data by
others must provide this same notification.}
}
\examples{
\dontrun{

inventory <- get_inventory()
}
}
\author{
Adam H Sparks, \email{adamhsparks@gmail.com}
}
