% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_history.R
\name{get_isd_history}
\alias{get_isd_history}
\title{Get the Most Recent isd_history File}
\usage{
get_isd_history()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Get the Most Recent isd_history File
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
get_isd_history()
\dontshow{\}) # examplesIf}
}
\seealso{
Other metadata: 
\code{\link{get_inventory}()},
\code{\link{get_updates}()}
}
\concept{metadata}
