% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_station_list}
\alias{get_station_list}
\title{Download the Latest Station List From the NCDC Server}
\usage{
get_station_list()
}
\value{
\code{\link[data.table]{data.table}} object of station metadata.
}
\description{
This function downloads the latest station list from the NCDC FTP server.
This list includes metadata for all stations including unique identifiers,
country, state (if in US), latitude, longitude, elevation and when weather
observations begin and end.  Stations with invalid latitude and longitude
values will not be included.
}
\examples{
\dontrun{
GSOD_stations <- get_station_list()
}
}
\author{
Adam Sparks, \email{adamhsparks@gmail.com}
}

