% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_updates.R
\name{get_updates}
\alias{get_updates}
\title{Get updates.txt With Information on Updates to the GSOD Data Set}
\usage{
get_updates()
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object
}
\description{
Gets and imports the 'updates.txt' file that has a change log of GSOD data.
Changes are shown in order from most recent to oldest changes by the "DATE"
field.  Column names follow \CRANpkg{GSODR} naming conventions.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
get_updates()
\dontshow{\}) # examplesIf}
}
\concept{metadata}
