% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\docType{package}
\name{GSODR-deprecated}
\alias{GSODR-deprecated}
\alias{get_station_list}
\alias{GSODR-deprecated-package}
\title{Deprecated function(s) in the GSODR package}
\usage{
get_station_list()
}
\description{
These functions are provided for compatibility with older version of
the GSODR package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{get_station_list} \tab now superceded by \code{\link{update_station_list}}\cr
}
\code{get_station_list} was used to download the list of weather stations
and the corresponding metadata. The station list is no longer fetched 
on-the-fly. Instead a version is supplied with the GSODR package upon
installation. To update the internal database, please use
\code{\link{update_station_list}}.
}

