% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SRTM_GSOD_elevation}
\alias{SRTM_GSOD_elevation}
\title{SRTM_GSOD_elevation}
\format{A data frame with 28339 observations of 2 elements:
\describe{
  \item{STNID}{Unique station ID, a concatenation of USAF and WBAN number,
  used for merging with station data weather files}
  \item{ELEV_M_SRTM_90m}{Elevation in metres extracted from SRTM data (Jarvis
  \emph{et al}. 2008)}
}

Station elevation information for the US National Centers for Environmental
Information (NCEI) Global Surface Summary of the Day (GSOD) weather data.
The original file has missing and incorrect geographic data including
location (LAT/LON) and elevation.  This data frame provides an alternative
set of elevation values with the following changes to the original list of
stations from the NCEI:
\enumerate{
\item{Stations with both a latitude and longitude of 0 degrees were removed.}
\item{Stations with longitude values that are beyond -180/180 degrees were
removed.}
\item{Stations with latitude values that are beyond -90/90 degrees were
removed.}
\item{A new field for elevation is included, ELEV_M_SRTM_90m. This was
created using mean values of a 200m buffer around the reported LAT/LON
station location within the CGIAR-CSI hole-filled 90m SRTM digital elevation
model (Jarvis \emph{et al}. 2008).}
}
For more on this, please consult the document available from the GSODR
GitHub repository detailing the process used to generate these data,
\url{https://github.com/ropensci/GSODR/blob/master/data-raw/fetch_isd-history.md}}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}
}
\usage{
SRTM_GSOD_elevation
}
\description{
SRTM_GSOD_elevation
}
\note{
Users of these data should take into account the following (from the
NCEI website): \dQuote{The following data and products may have conditions
placed on their international commercial use.  They can be used within the
U.S. or for non-commercial international activities without restriction. The
non-U.S. data cannot be redistributed for commercial purposes.
Re-distribution of these data by others must provide this same notification.}

The \code{SRTM_GSOD_elevation} data are automatically loaded with the
\code{GSODR} package and merged with the latest available data from the NCEI
in the "isd-history.csv" file.
}
\references{
{Jarvis, A., Reuter, H. I, Nelson, A., Guevara, E. (2008)
Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM
90m Database  \url{http://srtm.csi.cgiar.org}}
}
\keyword{datasets}
