% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory.R
\name{get_inventory}
\alias{get_inventory}
\title{Download and Return a data.table Object of GSOD Weather Station Data Inventories}
\usage{
get_inventory()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because the handling of country inventories has
changed and I did not want to introduce geospatial package dependencies for
this package.
}
\keyword{internal}
