% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory.R
\name{print.GSODR.Info}
\alias{print.GSODR.Info}
\title{Prints GSODR.info object}
\usage{
\method{print}{GSODR.Info}(x, ...)
}
\arguments{
\item{x}{GSODR.Info object}

\item{...}{ignored}
}
\description{
Prints GSODR.info object
}
