% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_station_list.R
\name{update_station_list}
\alias{update_station_list}
\title{Download Latest isd-history.csv File and Update an Internal Database}
\usage{
update_station_list()
}
\value{
Called for side-effects of saving a file to disk, returns an
invisible \code{NULL}.
}
\description{
This function downloads the latest station list (isd-history.csv) from the
\acronym{NCEI} server and updates the data distributed with \CRANpkg{GSODR}
to the latest stations available.  These data provide unique identifiers,
country, state (if in U.S.) and when weather observations begin and end.
}
\details{
Care should be taken when using this function if reproducibility is necessary
as different machines with the same version of \CRANpkg{GSODR} can end up
with different versions of the 'isd_history.csv' file internally.

There is no need to use this unless you know that a station exists in the
isd_history.csv file that is not available in the self-contained
database distributed with \CRANpkg{GSODR}.

To directly access these data, use: \cr
\code{load(system.file("extdata", "isd_history.rda", package = "GSODR"))}

To see the latest version available from the \acronym{NCEI} server, please
refer to \code{\link[=get_isd_history]{get_isd_history()}}.
}
\examples{
\dontrun{
update_station_list()
}
}
\seealso{
\code{\link[=get_isd_history]{get_isd_history()}}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
