% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_station_list.R
\name{update_station_list}
\alias{update_station_list}
\title{Download Latest isd-history.csv File and Update an Internal Database}
\usage{
update_station_list()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated because I realised that it broke reproducibility
and the handling of country inventories has changed and I did not want to
introduce geospatial package dependencies for this package.
}
\keyword{internal}
